Clean-up component
