+++ i18n/en.pot
@@ -5,8 +5,8 @@ msgstr ""
 "Content-Type: text/plain; charset=utf-8\n"
 "Content-Transfer-Encoding: 8bit\n"
 "Plural-Forms: nplurals=2; plural=(n != 1)\n"
-"POT-Creation-Date: 2021-10-28T10:02:24.060Z\n"
-"PO-Revision-Date: 2021-10-28T10:02:24.060Z\n"
+"POT-Creation-Date: 2021-11-09T19:21:27.238Z\n"
+"PO-Revision-Date: 2021-11-09T19:21:27.238Z\n"
 
 msgid "Data values - Create/update"
 msgstr ""
@@ -21,6 +21,12 @@ msgid "Unknown model"
 msgstr ""
 
 msgid "Metadata"
+msgstr ""
+
+msgid "Sign in with Microsoft"
+msgstr ""
+
+msgid "Unknown user"
 msgstr ""
 
 msgid "Add"
