+++ src/data/repositories/MetadataD2ApiRepository.ts
@@ -3,7 +3,11 @@ import { FutureData } from "../../domain/entities/Future";
 import { ImportResult, ImportStats } from "../../domain/entities/ImportResult";
 import { Instance } from "../../domain/entities/Instance";
 import { MetadataPayload } from "../../domain/entities/Metadata";
-import { ListMetadataResponse, ListMetadataOptions, MetadataRepository } from "../../domain/repositories/MetadataRepository";
+import {
+    ListMetadataResponse,
+    ListMetadataOptions,
+    MetadataRepository,
+} from "../../domain/repositories/MetadataRepository";
 import i18n from "../../locales";
 import { D2Api, D2ApiDefinition, MetadataResponse, Stats } from "../../types/d2-api";
 import { getD2APiFromInstance } from "../../utils/d2-api";
